Tests to demonstrate mirage association issues
I have a provider, schedule and location. A provider is linked to a
location through a schedule.

In the first test I create 10 providers and insure that the other
collections are empty.

In the second test I create a location, provider and schedule to link
the two. The associations are validate.

The last test is a repeat of the first. 50% of the time the tests fail
with `Mirage: Couldn't find provider with id = 2` when the provider
collection is accessed.